buffer: prevent 'time of check time of use' (CID 153006)

+++ src/flb_buffer_chunk.c
@@ -517,7 +517,10 @@ int flb_buffer_chunk_delete(struct flb_buffer_worker *worker,
         snprintf(path, sizeof(path) - 1, "%soutgoing/%s",
                  FLB_BUFFER_PATH(worker),
                  real_name);
-        unlink(path);
+        ret = stat(path, &st);
+        if (ret == 0 && S_ISREG(st.st_mode)) {
+            unlink(path);
+        }
     }
 
     free(target);
